Feature merge

+++ dmc/cleansing.py
@@ -46,7 +46,7 @@ def unproven_cleansing(df: pd.DataFrame) -> pd.DataFrame:
     # 0. (no voucher) is mostly appearing in data thus we can easily substitute
     df.voucherID = np.nan_to_num(df.voucherID).astype(np.int)
     # incorrect substitution for rrp, use imputation later
-    df.rrp = np.nan_to_num(df.rrp).astype(np.int)
+    df.rrp = np.nan_to_num(df.rrp).astype(np.float32)
     return df
